Refactor skill-review template to route component
