+++ mon-pix/app/routes/campaigns/assessment/skill-review.js
@@ -19,6 +19,6 @@ export default class SkillReviewRoute extends Route.extend(SecuredRouteMixin) {
     await model.campaignParticipation.belongsTo('campaignParticipationResult').reload();
     await model.campaignParticipation.belongsTo('campaign').reload({ include: 'targetProfile' });
     const improvableNextChallenge = await this.store.queryRecord('challenge', { assessmentId: model.assessment.id, tryImproving: true });
-    this.controllerFor('campaigns.assessment.skill-review').set('displayImprovementButton', !!improvableNextChallenge);
+    model.displayImprovementButton = !!improvableNextChallenge;
   }
 }
