+++ mon-pix/app/templates/campaigns/assessment/skill-review.hbs
@@ -1,127 +1,7 @@
 {{page-title (t 'pages.skill-review.title')}}
 
 <div class="background-banner-wrapper">
-
-  <div class="skill-review__banner">
-    <AssessmentBanner @title={{this.model.assessment.title}} @displayHomeLink={{false}} />
-  </div>
-
-  <PixBlock @shadow="heavy" class="skill-review__result-container">
-    {{#if this.model.campaignParticipation.campaign.isArchived}}
-
-      <div class="skill-review__campaign-archived">
-        <img class="skill-review__campaign-archived-image " src="{{this.rootURL}}/images/bees/fat-bee.svg" alt="" role="none">
-        <p class="skill-review__campaign-archived-text">
-          {{t 'pages.skill-review.archived' htmlSafe=true}}
-        </p>
-        <LinkTo @route="index" class="skill-review-share__back-to-home link">
-          {{t 'pages.skill-review.actions.continue'}}
-        </LinkTo>
-      </div>
-
-    {{else}}
-      <div class="skill-review__result-and-action">
-        <h2 class="sr-only">{{t 'pages.skill-review.abstract-title'}}</h2>
-        {{#if this.showStages}}
-          <ReachedStage
-            @stageCount={{this.stageCount}}
-            @starCount={{this.reachedStage.starCount}}
-            @percentage={{model.campaignParticipation.campaignParticipationResult.masteryPercentage}}
-            @imageUrl={{model.campaignParticipation.campaign.targetProfile.imageUrl}}
-          />
-        {{/if}}
-
-          <div class="skill-review__share">
-
-            {{#if this.showStages}}
-              <div class="skill-review-share__stage-congrats">
-                <div class="stage-congrats__title">
-                  {{this.reachedStage.title}}
-                </div>
-                <div class="stage-congrats__message">
-                  {{this.reachedStage.message}}
-                </div>
-              </div>
-            {{else}}
-              <p class="rounded-panel-title skill-review-result__abstract">
-                {{t 'pages.skill-review.abstract' percentage=this.model.campaignParticipation.campaignParticipationResult.masteryPercentage htmlSafe=true}}
-              </p>
-            {{/if}}
-            <h2 class="sr-only">{{t 'pages.skill-review.send-title'}}</h2>
-
-            <div class="skill-review-result__share-container {{if this.showStages "skill-review-result__share-container--left"}}">
-              {{#unless this.displayErrorMessage}}
-                <CampaignShareButton
-                  @isShared={{this.model.campaignParticipation.isShared}}
-                  @displayLoadingButton={{this.displayLoadingButton}}
-                  @shareCampaignParticipation={{this.shareCampaignParticipation}}
-                />
-              {{else}}
-                <div class="skill-review-share__error" aria-live="polite">
-                  {{t 'pages.skill-review.not-finished'}}
-                </div>
-                <LinkTo @route="campaigns.start-or-resume" @model={{this.model.campaignParticipation.campaign.code}}
-                        class="resume-campaign-banner__button button button--big button--link"><span class="sr-only">{{t
-                        'pages.profile.resume-campaign-banner.accessibility.resume'}}</span>{{t
-                        'pages.profile.resume-campaign-banner.actions.resume'}}</LinkTo>
-              {{/unless}}
-            </div>
-        </div>
-      </div>
-
-      {{#if this.showBadges}}
-        {{#unless this.hideBadgesTitle}}
-          <h2 class="skill-review-result__badge-subtitle">
-            {{t 'pages.skill-review.badges-title'}}
-          </h2>
-        {{/unless}}
-        <div class="badge-acquired-container">
-          {{#each this.acquiredBadges as |badge|}}
-            <BadgeAcquiredCard
-              @title={{badge.title}}
-              @message={{badge.message}}
-              @imageUrl={{badge.imageUrl}}
-              @altMessage={{badge.altMessage}}
-            />
-          {{/each}}
-        </div>
-        <div class="skill-review-result__dash-line"></div>
-      {{/if}}
-
-      {{#if this.displayImprovementButton}}
-        <SkillReviewTryAgain
-          @campaignParticipation={{this.model.campaignParticipation}}
-          @improvementCampaignParticipation={{this.improvementCampaignParticipation}}/>
-      {{/if}}
-    {{/if}}
-
-    <div class="skill-review-result__dash-line"></div>
-
-    <main>
-     <div class="skill-review-result__table-header">
-        <h2 class="skill-review-result__subtitle">
-          {{t 'pages.skill-review.details.title'}}
-        </h2>
-        <CircleChart @value={{this.model.campaignParticipation.campaignParticipationResult.masteryPercentage}}>
-          <span aria-label="{{t "pages.skill-review.details.result"}}" class="skill-review-table-header__circle-chart-value">
-            {{@model.campaignParticipation.campaignParticipationResult.masteryPercentage}}%
-          </span>
-        </CircleChart>
-      </div>
-
-     <CampaignSkillReviewCompetenceResult
-       @showCleaCompetences={{this.showCleaCompetences}}
-       @competenceResults={{this.model.campaignParticipation.campaignParticipationResult.competenceResults}}
-       @partnerCompetenceResults={{this.model.campaignParticipation.campaignParticipationResult.cleaBadge.partnerCompetenceResults}}
-     />
-    </main>
-
-    <div class="skill-review-result__information">
-      <FaIcon @icon='info-circle' class='skill-review-information__info-icon' aria-hidden="true"/>
-      <div class="skill-review-information__text">
-        {{t 'pages.skill-review.information'}}
-      </div>
-    </div>
-
-  </PixBlock>
+  <Routes::Campaigns::Assessment::SkillReview
+          @model={{@model}}
+  />
 </div>
